Move glue to submodule
