+++ setup.py
@@ -45,6 +45,7 @@ setup(
         "docutils>=0.15",
         "sphinx>=2,<3",
         "jupyter_sphinx==0.2.4a1",
+        "ipython",
         "nbformat",
         "nbconvert",
         "pyyaml",
@@ -58,11 +59,9 @@ setup(
             "pytest-cov",
             "pytest-regressions",
             "beautifulsoup4",
-            "ipython",
         ],
         "rtd": [
             "sphinxcontrib-bibtex",
-            "ipython",
             "ipywidgets",
             "pandas",
             "numpy",
