+++ tests/test_glue.py
@@ -6,7 +6,7 @@ from IPython.core.displaypub import DisplayPublisher
 from docutils.parsers.rst import directives, roles
 from docutils.transforms import Transformer
 
-from myst_nb import glue
+from myst_nb.glue import glue, sphinx, utils
 from myst_nb.parser import NotebookParser
 from myst_nb.transform import CellOutputsToNodes
 
@@ -33,15 +33,15 @@ def mock_ipython():
 def patch_docutils():
     _directives = copy(directives._directives)
     _roles = copy(roles._roles)
-    directives._directives["paste"] = glue.Paste
-    roles._roles["paste"] = glue.paste_role
+    directives._directives["paste"] = sphinx.Paste
+    roles._roles["paste"] = sphinx.paste_role
     yield None
     directives._directives = _directives
     roles._roles = _roles
 
 
 def test_glue_func_text(mock_ipython):
-    glue.glue("a", "b")
+    glue("a", "b")
     assert mock_ipython.publish_calls == [
         {
             "metadata": {"scrapbook": {"name": "a", "mime_prefix": ""}},
@@ -58,7 +58,7 @@ def test_glue_func_obj(mock_ipython):
         def _repr_html_(self):
             return "<p>repr</p>"
 
-    glue.glue("a", Obj())
+    glue("a", Obj())
     assert mock_ipython.publish_calls == [
         {
             "metadata": {"scrapbook": {"name": "a", "mime_prefix": ""}},
@@ -75,7 +75,7 @@ def test_glue_func_obj_no_display(mock_ipython):
         def _repr_html_(self):
             return "<p>repr</p>"
 
-    glue.glue("a", Obj(), display=False)
+    glue("a", Obj(), display=False)
     assert mock_ipython.publish_calls == [
         {
             "metadata": {
@@ -94,15 +94,15 @@ def test_glue_func_obj_no_display(mock_ipython):
 
 def test_find_glued_key(get_notebook):
 
-    bundle = glue.find_glued_key(get_notebook("with_glue.ipynb"), "key_text1")
+    bundle = utils.find_glued_key(get_notebook("with_glue.ipynb"), "key_text1")
     assert bundle == {"key_text1": "'text1'"}
 
     with pytest.raises(KeyError):
-        glue.find_glued_key(get_notebook("with_glue.ipynb"), "unknown")
+        utils.find_glued_key(get_notebook("with_glue.ipynb"), "unknown")
 
 
 def test_find_all_keys(get_notebook):
-    keys = glue.find_all_keys(get_notebook("with_glue.ipynb"))
+    keys = utils.find_all_keys(get_notebook("with_glue.ipynb"))
     assert set(keys) == {
         "key_text1",
         "key_float",
@@ -117,7 +117,7 @@ def test_parser(patch_docutils, mock_document_in_temp, get_notebook, file_regres
     parser.parse(get_notebook("with_glue.ipynb").read_text(), mock_document_in_temp)
 
     transformer = Transformer(mock_document_in_temp)
-    transformer.add_transforms([CellOutputsToNodes, glue.PasteNodesToDocutils])
+    transformer.add_transforms([CellOutputsToNodes, sphinx.PasteNodesToDocutils])
     transformer.apply_transforms()
 
     file_regression.check(mock_document_in_temp.pformat(), extension=".xml")
